extraction/requirements-extract.txt
spacy==3.7.2
pyyaml==6.0.1
